ai/requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
clip-interrogator==0.6.0
colorama==0.4.6
contourpy==1.3.0
controlnet-aux==0.0.3
cycler==0.12.1
diffusers==0.27.2
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.108.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gradio==4.21.0
gradio_client==0.12.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
IOPaint==1.4.3
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
narwhals==1.6.3
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.26.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.7.1
piexif==1.1.3
Pillow==9.5.0
psutil==6.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.7.2
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
ruff==0.6.4
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.32.0.post1
sympy==1.13.2
tifffile==2024.8.30
timm==1.0.9
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typer-config==1.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.30.6
wcwidth==0.2.13
websockets==11.0.3
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
yacs==0.1.8
yarl==1.9.4
zipp==3.20.1
nvidia-ml-py3==7.352.0
scikit-image==0.24.0